old/bio-project/requirements.txt
openai>=1.40.0
python-dotenv>=1.0.1
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9

